fix viewed car commponents, moved them to the same folder
