+++ frontend/src/components/viewed-cars/components/short-car-card/short-car-card.tsx
