+++ frontend/src/components/viewed-cars/components/short-car-card/styles.module.scss
@@ -31,7 +31,7 @@
   margin-top: 0;
   margin-bottom: 5px;
   font-weight: 800;
-  font-size: 18px;
+  font-size: 1.125rem;
   text-transform: uppercase;
 }
