+++ frontend/src/components/viewed-cars/components/viewed-cars-grid/style.module.scss
