+++ frontend/src/components/viewed-cars/styles.module.scss
@@ -7,7 +7,7 @@
   &Title {
     margin-top: 0;
     font-weight: 800;
-    font-size: 22;
+    font-size: 1.375rem;
     text-align: left;
     text-transform: uppercase;
   }
